Removing XSPH Correction + New ST model
Removing the XSPH correction as it appears to introduce instability in the 3D code, and its also not necessary for the use application. Also of significant note is the option of another surface tension model, based off of He et al (2014). It claims to be better for thin features, and this may be so but it appears to be a bit more noisy in comparison, though I need to do more comparison. Little information is also available on the factor kappa in the model, and its value for fluids.

+++ src/3D/Neighbours.h
@@ -12,19 +12,22 @@ void FindNeighbours(Sim_Tree &NP1_INDEX, FLUID &fvar, State &pnp1,outl &outlist)
 	nanoflann::SearchParams params;
 	outlist.erase(outlist.begin(),outlist.end());
 	ldouble search_radius = fvar.sr;
+
 	/*Find neighbour list*/
-	for(unsigned int i=0; i<pnp1.size(); ++i)
+	for(uint i=0; i<pnp1.size(); ++i)
 	{
+		// std::cout << pnp1[i].list.size();
 		std::vector<std::pair<size_t, ldouble>> matches; /* Nearest Neighbour Search*/
-
 		NP1_INDEX.index->radiusSearch(&pnp1[i].xi[0], search_radius, matches, params);
-
-		std::vector<unsigned int> temp;
+		
+		std::vector<uint> temp;
 		for (auto &j:matches)
 		{
-			temp.emplace_back(j.first);
+			temp.emplace_back(static_cast<uint>(j.first));
 		}
 		outlist.emplace_back(temp);
+
+		// std::cout << "  " << pnp1[i].list.size() << std::endl;
 	}
 }
